feat: add expected swap amount calculation
