+++ package.json
@@ -18,7 +18,7 @@
     "@dcspark/cardano-multiplatform-lib-nodejs": "^3.1.0"
   },
   "dependencies": {
-    "@wingriders/dex-serializer": "^0.1.3",
+    "@wingriders/dex-serializer": "^0.1.4",
     "axios": "^1.1.3",
     "sha3": "^2.1.4"
   },
