+++ src/constants.ts
@@ -3,3 +3,10 @@ export const LIQUIDITY_POOL_VALIDITY_ASSET: string = `${LIQUIDITY_POLICY_ID}4c`;
 
 export const LIQUIDITY_POOL_MIN_ADA = "3000000";
 export const INT_MAX_64 = "9223372036854775807";
+
+export const SWAP_FEE_IN_BASIS = "35";
+
+/* NOTE move these to dex-serializer and split based on network */
+export const REQUEST_SCRIPT_HASH = "86ae9eebd8b97944a45201e4aec1330a72291af2d071644bba015959";
+export const REQUEST_BATCHER_FEE = "2000000";
+export const REQUEST_OIL = "2000000";
